Grab jenkins config with auth bits from private gs bucket
This grabs the jenkins config.xml file including the authentication
bits from a project-private google storage bucket so a frestly created
instance has authentication with github set up out of the box. This
means that you have to be authenticated with cloud and a member of the
google compute project in order to set up a new image.